requirements.txt
flask
python-dotenv
pymongo[srv]

gunicorn